將各個小功能拆分成各個啟動圖示
