+++ app/src/main/java/com/example/jack8/floatwindow/CalculatoLauncher.java
@@ -8,16 +8,16 @@ import android.support.annotation.RequiresApi;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+public class CalculatoLauncher extends AppCompatActivity {
 
-public class MainActivity extends AppCompatActivity {
     @RequiresApi(api = Build.VERSION_CODES.M)
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         //setContentView(R.layout.window);
-        if (Build.VERSION.SDK_INT>=Build.VERSION_CODES.M&&!Settings.canDrawOverlays(MainActivity.this))
+        if (Build.VERSION.SDK_INT>=Build.VERSION_CODES.M&&!Settings.canDrawOverlays(CalculatoLauncher.this))
             startActivityForResult(new Intent(Settings.ACTION_MANAGE_OVERLAY_PERMISSION,
-                    Uri.parse("package:" + MainActivity.this.getPackageName())), 1);
+                    Uri.parse("package:" + CalculatoLauncher.this.getPackageName())), 1);
         else {
             startFloatWindow();
         }
@@ -25,7 +25,7 @@ public class MainActivity extends AppCompatActivity {
     private void startFloatWindow(){
         Intent intent=new Intent(this, FloatServer.class);
 
-        intent.putExtra("intent",FloatServer.OPEN_MAIN_MENU);
+        intent.putExtra("intent",FloatServer.OPEN_CALCULATO);
 
         if(Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
             startService(intent);
