+++ app/src/main/java/com/example/jack8/floatwindow/WebBrowserLauncher.java
@@ -8,24 +8,34 @@ import android.support.annotation.RequiresApi;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+public class WebBrowserLauncher extends AppCompatActivity {
 
-public class MainActivity extends AppCompatActivity {
     @RequiresApi(api = Build.VERSION_CODES.M)
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         //setContentView(R.layout.window);
-        if (Build.VERSION.SDK_INT>=Build.VERSION_CODES.M&&!Settings.canDrawOverlays(MainActivity.this))
+        if (Build.VERSION.SDK_INT>=Build.VERSION_CODES.M&&!Settings.canDrawOverlays(WebBrowserLauncher.this))
             startActivityForResult(new Intent(Settings.ACTION_MANAGE_OVERLAY_PERMISSION,
-                    Uri.parse("package:" + MainActivity.this.getPackageName())), 1);
+                    Uri.parse("package:" + WebBrowserLauncher.this.getPackageName())), 1);
         else {
             startFloatWindow();
         }
     }
     private void startFloatWindow(){
         Intent intent=new Intent(this, FloatServer.class);
-
-        intent.putExtra("intent",FloatServer.OPEN_MAIN_MENU);
+        Intent extra_intent = getIntent();
+        String url = null;
+        if(extra_intent.getStringExtra(Intent.EXTRA_TEXT) != null)//帶有外部字串的啟動方法
+            url = extra_intent.getStringExtra(Intent.EXTRA_TEXT);
+        else if(extra_intent.getDataString() != null)//呼叫瀏覽器的的啟動方式
+            url = extra_intent.getDataString();
+        if(url == null) {
+            intent.putExtra("intent",FloatServer.OPEN_WEB_BROWSER);
+        }else{
+            intent.putExtra("intent",FloatServer.OPEN_WEB_BROWSER | FloatServer.OPEN_EXTRA_URL);
+            intent.putExtra("extra_url",url);
+        }
 
         if(Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
             startService(intent);
