+++ app/src/main/res/values/strings.xml
@@ -68,4 +68,5 @@
     <string name="game_over">game over</string>
     <string name="brick">brick</string>
     <string name="enable_ads_block">Enable ads block</string>
+    <string name="calculato">Calculato</string>
 </resources>
